--- de.metas.acct.base/src/main/java/de/metas/acct/impl/PlainFactAcctLogDAO.java
@@ -1,37 +1,0 @@
-package de.metas.acct.impl;
-
-import org.slf4j.Logger;
-import de.metas.logging.LogManager;
-
-/*
- * #%L
- * de.metas.acct.base
- * %%
- * Copyright (C) 2016 metas GmbH
- * %%
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as
- * published by the Free Software Foundation, either version 2 of the
- * License, or (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
- * <http://www.gnu.org/licenses/gpl-2.0.html>.
- * #L%
- */
-
-public class PlainFactAcctLogDAO extends FactAcctLogDAO
-{
-	private static final transient Logger logger = LogManager.getLogger(PlainFactAcctLogDAO.class);
-
-	@Override
-	public void updateFactAcctEndingBalanceForTag(String processingTag)
-	{
-		logger.warn("Updating Fact_Acct_EndingBalance not supported. Skipped.");
-	}
-}
