+++ de.metas.acct.base/src/main/sql/postgresql/system/43-de.metas.acct/5497390_sys_gh4345_rename.sql
@@ -1,0 +1,5 @@
+-- 2018-07-11T13:42:50.230
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE C_Queue_PackageProcessor SET Classname='de.metas.acct.aggregation.async.FactAcctLogWorkpackageProcessor',Updated=TO_TIMESTAMP('2018-07-11 13:42:50','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE C_Queue_PackageProcessor_ID=540035
+;
+
